fix(rtx): preserve dim-light mass in the geometry-light pick CDF
The Light Pick CDF was a normalized cumulative float, so a light whose power is
below float epsilon of the total quantized to a zero-width interval: the
NEE-sample side (inverseSampleCDF, instancePickProbability delta) could never
select it, yet the hit-side geometryLightHitPdf still reported a positive
analytic pick probability. The deposit's MIS weight then down-weighted the
forward path for a light next-event never delivered — the dim emitter rendered
too dark in a bright+dim scene.

Store and accumulate the CDF in double (masses preserved to ~1e-16 of total) and
difference the pick-probability delta in double, so any physically meaningful
light stays selectable with a delta that matches the hit-side pNee. The float
HDRI/primitive CDFs are untouched: inverseSampleCDF is now templated on the CDF
element type.

Normalize by the double cumulative total, not the float m_totalLightPower:
dividing by the float total can push the last entry above 1.0 when float dropped
a dim light's mass, leaving trailing dim lights unselectable while their pNee
stays positive. The double total makes the last entry exactly 1.0 and preserves
those masses.

No test regression (the effect is below the float-epsilon noise floor to
reproduce automatically); this is precision hardening.

+++ devices/rtx/device/gpu/gpu_objects.h
@@ -507,8 +507,8 @@ struct MaterialGPUData
   bool emissionIsConstant{false};
 
   // Emission is not provably zero (constant, sampler, or attribute bound). The
-  // hit-side Geometry Light MIS gate; kept in sync by the material's own commit,
-  // so it is never stale.
+  // hit-side Geometry Light MIS gate; kept in sync by the material's own
+  // commit, so it is never stale.
   bool emissionIsSampleable{false};
 
   // Mean emitted radiance, sizing the Geometry Light Pick Power on both the
@@ -808,8 +808,8 @@ struct InstanceLightGPUData
   mat4 xfm; // Transform for this light instance
   // For a Geometry Light: index into WorldGPUData::surfaceInstances[] of the
   // surface instance it was synthesized from, so the NEE sampler can evaluate
-  // emission against the REAL instance (instance-uniform attributes resolve like
-  // the path-hit deposit). -1 for authored/HDRI lights.
+  // emission against the REAL instance (instance-uniform attributes resolve
+  // like the path-hit deposit). -1 for authored/HDRI lights.
   DeviceObjectIndex surfaceInstanceIndex = -1;
 };
 
@@ -835,7 +835,10 @@ struct WorldGPUData
   // Normalized cumulative Pick Power over lightInstances (length
   // numLightInstances, last entry == 1); pick a slot with inverseSampleCDF. The
   // slot's discrete pick probability is lightPickCdf[i]-lightPickCdf[i-1].
-  const float *lightPickCdf;
+  // Double so a dim light's mass (below float epsilon of the total) is
+  // preserved as a nonzero interval — else it is unselectable while the
+  // hit-side pNee is still positive, biasing the deposit's MIS weight.
+  const double *lightPickCdf;
   float totalLightPower; // sum of un-normalized instance Pick Powers
   float hdriPower; // subset sum over HDRI instances (env-MIS pick probability)
   float sceneRadius; // bounding-sphere radius, for the ambient term's power
