+++ devices/rtx/device/world/World.h
@@ -115,7 +115,7 @@ struct World : public Object
   HostDeviceArray<InstanceLightGPUData> m_instanceHdriLightGPUData;
 
   // Power-proportional Light Pick, rebuilt with the light instances.
-  HostDeviceArray<float> m_lightPickCdf;
+  HostDeviceArray<double> m_lightPickCdf;
   float m_totalLightPower{0.f};
   float m_hdriPower{0.f};
   float m_sceneRadius{0.f};
